kb/requirements.txt
supabase
python-dotenv
ollama
numpy
beautifulsoup4
markdown
PyPDF2
unstructured
